integrated main branch

+++ main/include/wifi.h
@@ -3,5 +3,5 @@
 #define __WIFI_H__
 
 void wifi_init_sta(void);
-void send_data(uint32_t* temperature, uint32_t* humidity, uint32_t* pressure);
+void send_data(double* temperature, double* humidity, double* pressure);
 #endif  // __WIFI_H__
